+++ .gitignore
@@ -2,6 +2,8 @@
 Thumbs.db
 ehthumbs.db
 
+.vs/
+
 # Folder config file
 Desktop.ini
 
@@ -45,3 +47,6 @@ $RECYCLE.BIN/
 Network Trash Folder
 Temporary Items
 .apdisk
+
+bin/
+obj/
