--- XamarinCircleButton/MainActivity.cs
@@ -1,19 +1,0 @@
-﻿using Android.App;
-using Android.Widget;
-using Android.OS;
-
-namespace XamarinCircleButton
-{
-    [Activity(Label = "XamarinCircleButton", MainLauncher = true, Icon = "@drawable/icon")]
-    public class MainActivity : Activity
-    {
-        protected override void OnCreate(Bundle bundle)
-        {
-            base.OnCreate(bundle);
-
-            // Set our view from the "main" layout resource
-            // SetContentView (Resource.Layout.Main);
-        }
-    }
-}
-
