--- XamarinCircleButton/Properties/AndroidManifest.xml
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="XamarinCircleButton.XamarinCircleButton" android:versionCode="1" android:versionName="1.0">
-  <uses-sdk android:minSdkVersion="16" />
-  <application android:label="XamarinCircleButton"></application>
-</manifest>
