+++ qs/CircleButtonQs/MainActivity.cs
@@ -1,0 +1,17 @@
+﻿using Android.App;
+using Android.OS;
+
+namespace CircleButtonQs
+{
+    [Activity(Label = "@string/app_name", MainLauncher = true)]
+    public class MainActivity : Activity
+    {
+
+        protected override void OnCreate(Bundle savedInstanceState)
+        {
+            base.OnCreate(savedInstanceState);
+            SetContentView(Resource.Layout.activity_main);
+        }
+    }
+}
+
