+++ qs/CircleButtonQs/Resources/layout/activity_main.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<FrameLayout
+	xmlns:android="http://schemas.android.com/apk/res/android"
+	xmlns:app="http://schemas.android.com/apk/res-auto"
+	android:layout_width="match_parent"
+	android:layout_height="match_parent">
+
+	<at.markushi.ui.CircleButton
+		android:layout_width="64dip"
+		android:layout_height="64dip"
+		android:src="@drawable/ic_action_tick"
+		app:cb_color="#99CC00"
+		app:cb_pressedRingWidth="8dip" />
+
+</FrameLayout>
+
