+++ qs/CircleButtonQs/Resources/mipmap-anydpi-v26/ic_launcher_round.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<adaptive-icon
+    xmlns:android="http://schemas.android.com/apk/res/android">
+    <background
+        android:drawable="@color/ic_launcher_background" />
+    <foreground
+        android:drawable="@mipmap/ic_launcher_foreground" />
+</adaptive-icon>
