+++ qs/CircleButtonQs/Resources/values/colors.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <color
+        name="colorPrimary">#2c3e50</color>
+    <color
+        name="colorPrimaryDark">#1B3147</color>
+    <color
+        name="colorAccent">#3498db</color>
+</resources>
