+++ qs/CircleButtonQs/Resources/values/dimens.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<resources>
+    <dimen
+        name="fab_margin">16dp</dimen>
+</resources>
