+++ qs/CircleButtonQs/Resources/values/ic_launcher_background.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <color
+        name="ic_launcher_background">#2C3E50</color>
+</resources>
