+++ qs/CircleButtonQs/Resources/values/strings.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<resources>
+    <string
+        name="app_name">CircleButtonQs</string>
+    <string
+        name="action_settings">Settings</string>
+</resources>
