+++ qs/CircleButtonQs/Resources/values/styles.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<resources>
+    
+</resources>
